+++ src/App.js
@@ -2,7 +2,7 @@ import { Console } from '@woowacourse/mission-utils';
 import tryMatch from './modules/tryMatch';
 import getRandomUniqueNumber from './modules/getRandomUniqueNumber';
 import getNextStep from './modules/getNextStep';
-import Constant from './modules/Constant';
+import Constant from './constant/Constant';
 
 const { EXIT, START_PROMPT, EXIT_PROMPT } = Constant;
